requirements.txt
frontmatter
httpwatcher
python-slugify
markdown
jinja2